Create Word Game
This is a simple game for 2 players. It is coded in C++ and runs best on latest C++ compilers. Both the players have to remember all the words added in sequence. A player has tell them correctly in sequence, and then have to add a word to it thus making it more difficult to remember. Both player take turns to remember the sequence and add a word. The first person to tell an incorrect sequence loses.